requirements.txt
openai-whisper
gspread
google-auth
google-api-python-client
python-dotenv
requests
logger